Fix incorrect prototype for check_current_fs() after OpenBSD fix

+++ bacula/src/findlib/protos.h
@@ -84,7 +84,7 @@ bool path_list_add(JCR *jcr, uint32_t len, char *fname);
 bool fstype(FF_PKT *ff_pkt, char *fs, int fslen);
 bool fstype(char *fname, FF_PKT *ff_pkt, char *fs, int fslen);
 bool check_current_fs(char *fname, FF_PKT *ff, const char *fstype_name);
-bool check_current_fs(char *fname, FF_PKT *ff, uint32_t fstype_magic);
+bool check_current_fs(char *fname, FF_PKT *ff, uint64_t fstype_magic);
 
 /* from drivetype.c */
 bool drivetype(const char *fname, char *fs, int fslen);
